+++ src/OrderItem.php
@@ -74,7 +74,7 @@ class OrderItem extends Data
     /**
      * Mapping: Ingenico's field => OrderItem's field
      */
-    static $map = [
+    private static $map = [
         'ITEMATTRIBUTES' => self::ITEM_ATTRIBUTES,
         'ITEMCATEGORY' => self::ITEM_CATEGORY,
         'ITEMCOMMENTS' => self::ITEM_COMMENTS,
@@ -151,6 +151,7 @@ class OrderItem extends Data
                         break;
                     case self::ITEM_VAT_INCLUDED:
                         $result[$key] = !in_array((int) $result[$key], [0, 1]) ? 0 : $result[$key];
+                        break;
                     default:
                         // no break
                 }